+++ 2libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: ochristm <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/08 13:50:41 by ochristm          #+#    #+#             */
-/*   Updated: 2019/09/08 13:53:15 by ochristm         ###   ########.fr       */
+/*   Updated: 2019/09/21 15:45:23 by ochristm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char const *s, int fd)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s)
