+++ 2libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: ochristm <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/05 21:05:54 by ochristm          #+#    #+#             */
-/*   Updated: 2019/09/09 19:14:19 by ochristm         ###   ########.fr       */
+/*   Updated: 2019/09/21 15:36:07 by ochristm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *src)
 {
 	char	*str;
-	int		len;
+	size_t	len;
 
 	len = 0;
 	while (src[len])
