+++ 2libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ochristm <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/07 19:41:04 by ochristm          #+#    #+#             */
-/*   Updated: 2019/09/14 12:21:10 by ochristm         ###   ########.fr       */
+/*   Updated: 2019/09/21 15:48:43 by ochristm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s1 == NULL && s2 == NULL)
 		return (0);
-	while (s1[i] == s2[i] && i < (int)n && s1[i] != '\0')
+	while (s1[i] == s2[i] && i < n && s1[i] != '\0')
 		i++;
-	if (i == (int)n)
+	if (i == n)
 		return ((unsigned char)s1[i - 1] - (unsigned char)s2[i - 1]);
 	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 }
