+++ 2libft/ft_strstr.c
@@ -6,7 +6,7 @@
 /*   By: ochristm <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/07 19:30:33 by ochristm          #+#    #+#             */
-/*   Updated: 2019/09/09 19:22:43 by ochristm         ###   ########.fr       */
+/*   Updated: 2019/09/21 15:25:43 by ochristm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strstr(const char *big, const char *small)
 {
-	int i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
